Remove last dbmanager extension, consolidate everything in libraryservice
