+++ BookPlayer/Coordinators/FolderListCoordinator.swift
@@ -18,7 +18,6 @@ class FolderListCoordinator: ItemListCoordinator {
     folder: Folder,
     playerManager: PlayerManagerProtocol,
     importManager: ImportManager,
-    dataManager: DataManager,
     libraryService: LibraryServiceProtocol
   ) {
     self.folder = folder
@@ -28,7 +27,6 @@ class FolderListCoordinator: ItemListCoordinator {
       library: library,
       playerManager: playerManager,
       importManager: importManager,
-      dataManager: dataManager,
       libraryService: libraryService
     )
   }
@@ -38,7 +36,6 @@ class FolderListCoordinator: ItemListCoordinator {
     let viewModel = ItemListViewModel(folder: self.folder,
                                       library: self.library,
                                       playerManager: self.playerManager,
-                                      dataManager: self.dataManager,
                                       libraryService: self.libraryService,
                                       themeAccent: ThemeManager.shared.currentTheme.linkColor)
     viewModel.coordinator = self
